Add helper method to flush segments
