+++ src/include/duckdb/storage/compression/alp/alp_compress.hpp
@@ -225,7 +225,7 @@ public:
 		// Store the offset to the end of metadata (to be used as a backwards pointer in decoding)
 		Store<uint32_t>(NumericCast<uint32_t>(total_segment_size), dataptr);
 
-		checkpoint_state.FlushSegment(std::move(current_segment), std::move(handle), total_segment_size);
+		FlushCurrentSegment(total_segment_size);
 		data_bytes_used = 0;
 		vectors_flushed = 0;
 	}
