+++ src/storage/compression/dictionary/compression.cpp
@@ -103,9 +103,8 @@ bool DictionaryCompressionCompressState::CalculateSpaceRequirements(bool new_str
 
 void DictionaryCompressionCompressState::Flush(bool final) {
 	auto segment_size = Finalize();
-	auto &state = checkpoint_data.GetCheckpointState();
 	stats_writer.Merge(current_segment->GetStatsMutable());
-	state.FlushSegment(std::move(current_segment), std::move(handle), segment_size);
+	FlushCurrentSegment(segment_size);
 
 	if (!final) {
 		CreateEmptySegment();
